Revised "random" moves for small span
May not need multiple strategies; need to do a bit more testing, but
it seems to be working well at least for one or two small spans.

Also wound up redoing the acceleration algorithms; much simplified.
Skipping no longer adds steps but allows acceleration to increase if
laser is blocked by tape. Once a random move ends, if it is blocked
by tape, then it will create a move to the next unblocked area instead
of a random move.